Create the interface for the subsystem wrapper
