+++ consensus/src/detail/error.rs
