+++ consensus/src/detail/pow/mod.rs
