+++ consensus/src/detail/pow/temp.rs
